i_ot/requirements.txt
aniso8601==6.0.0
certifi==2019.6.16
chardet==3.0.4
coverage==4.5.3
dj-database-url==0.5.0
Django==2.2.4
django-crispy-forms==1.7.2
django-filter==2.0.0
django-graphql-jwt==0.2.1
django-redis==4.10.0
django-rest-registration==0.5.1
djangorestframework==3.9.4
djangorestframework-simplejwt==4.3.0
graphene==2.1.7
graphene-django==2.2.0
graphql-core==2.2.1
graphql-relay==2.0.0
idna==2.8
promise==2.2.1
psycopg2-binary==2.8.3
PyJWT==1.7.1
python-decouple==3.1
pytz==2019.2
redis==3.3.8
requests==2.22.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.3
whitenoise==4.1.3
